Move memlet propagation to dace.sdfg
