+++ dace/__init__.py
@@ -8,9 +8,9 @@ from .frontend.operations import *
 
 from .config import Config
 from .sdfg import SDFG, SDFGState, InterstateEdge
+from .sdfg.propagation import propagate_memlets_sdfg, propagate_memlet
 from .memlet import Memlet, EmptyMemlet
 from .graph import nodes
-from .graph.labeling import propagate_labels_sdfg, propagate_memlet
 from .symbolic import symbol
 
 # Run Jupyter notebook code
