+++ dace/transformation/dataflow/merge_arrays.py
@@ -2,7 +2,7 @@ from dace.transformation import pattern_matching
 from dace import memlet, registry
 from dace.graph import nodes
 from dace.sdfg import SDFGState
-from dace.graph.labeling import propagate_memlet
+from dace.sdfg.propagation import propagate_memlet
 
 
 @registry.autoregister_params(singlestate=True, strict=True)
