+++ doc/source/dace.graph.rst
@@ -4,14 +4,6 @@ dace.graph package
 Submodules
 ----------
 
-dace.graph.dot module
----------------------
-
-.. automodule:: dace.graph.dot
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 dace.graph.graph module
 -----------------------
 
@@ -20,26 +12,10 @@ dace.graph.graph module
    :undoc-members:
    :show-inheritance:
 
-dace.graph.labeling module
---------------------------
-
-.. automodule:: dace.graph.labeling
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 dace.graph.nodes module
 -----------------------
 
 .. automodule:: dace.graph.nodes
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-dace.sdfg.utils module
-------------------------
-
-.. automodule:: dace.sdfg.utils
    :members:
    :undoc-members:
    :show-inheritance:
