+++ doc/source/dace.sdfg.rst
@@ -1,0 +1,55 @@
+dace.sdfg package
+==================
+
+Submodules
+----------
+
+dace.sdfg.propagation module
+----------------------------
+
+.. automodule:: dace.sdfg.propagation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+dace.sdfg.scope module
+----------------------
+
+.. automodule:: dace.sdfg.scope
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+dace.sdfg.sdfg module
+---------------------
+
+.. automodule:: dace.sdfg.sdfg
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+dace.sdfg.utils module
+----------------------
+
+.. automodule:: dace.sdfg.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+dace.sdfg.validation module
+---------------------------
+
+.. automodule:: dace.sdfg.validation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: dace.sdfg
+   :members:
+   :undoc-members:
+   :show-inheritance:
