+++ runsdfg.py
@@ -1,12 +1,11 @@
 import dace
-import dace.graph.labeling
 import sys
 import time
 
 print(time.time(), 'loading')
 a = dace.SDFG.from_file(sys.argv[1])
 print(time.time(), 'propagating')
-dace.graph.labeling.propagate_labels_sdfg(a)
+dace.propagate_memlets_sdfg(a)
 print(time.time(), 'strict transformations')
 a.apply_strict_transformations()
 print(time.time(), 'saving')
